requirements.txt
pytest==8.2.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
Requests==2.31.0
Telethon==1.35.0
